rename: 대문자 i를 소문자자i로 바꾸기위한 tmp
